#110 Refactor and style the components
