+++ package.json
@@ -26,6 +26,7 @@
         "@types/react": "^18.2.6",
         "@types/react-dom": "^18.2.4",
         "@vitejs/plugin-react": "^4.0.0",
+        "sass": "^1.63.4",
         "typescript": "^5.0.4",
         "vite": "^4.3.9"
     }
