+++ src/App.tsx
@@ -1,13 +1,32 @@
-import { Container } from './Container';
-import { defaultProperty } from './properties';
-import './index.css';
+import { Container } from "./components";
+import "./style.scss";
+import githubMark from "./assets/github-mark.svg";
 
 
 function App() {
     return (
         <>
-            <h1>Simple Logo Generator</h1>
-            <Container defaults={defaultProperty} />
+            <div className="flex">
+                <h1>Simple Logo Generator</h1>
+            </div>
+            <Container />
+            <div className="contribution flex">
+                <div>
+                    <span className="small">
+                        If you've found any defects, please report it at
+                    </span>
+                    <br />
+                    <span className="small">
+                        If you're willing to contribute this project, visit
+                    </span>
+                </div>
+                <a
+                    href="https://github.com/KaoruNishikawa/SimpleLogoGenerator"
+                    target="_blank"
+                >
+                    <img className="git-link" src={githubMark} />
+                </a>
+            </div>
         </>
     );
 }
