--- src/Input.tsx
@@ -1,12 +1,0 @@
-function Input(props: { unit?: string, [key: string]: any }) {
-    const { unit, ...rest } = props;
-    return (
-        <div>
-            <input {...rest} />
-            {unit && <span>{unit}</span>}
-        </div>
-    );
-}
-
-
-export { Input };
