+++ src/components/AngleInput.scss
@@ -1,0 +1,50 @@
+.angle-input {
+    --disk-radius: 24px;
+    --pivot-radius: 16px;
+    --border-color: #ccc;
+
+    &:focus {
+        outline: none;
+        --border-color: #77f;
+    }
+
+    & input {
+        opacity: 0;
+        width: 0;
+        height: 0;
+    }
+
+    background-color: #eef;
+    border: 1px solid var(--border-color);
+    box-shadow: 2px 2px 2px #888;
+    border-radius: 50%;
+    width: calc(2 * var(--disk-radius));
+    height: calc(2 * var(--disk-radius));
+    position: relative;
+    top: 0;
+    left: 0;
+
+    & .pivot {
+        display: inline-block;
+        background-color: transparent;
+        border-radius: 50%;
+        width: calc(2 * var(--disk-radius));
+        height: calc(2 * var(--disk-radius));
+        position: absolute;
+        top: calc(var(--disk-radius) - var(--pivot-radius) - 7px);
+        left: calc(var(--disk-radius) - var(--pivot-radius) - 7px);
+
+        &::before {
+            position: absolute;
+            content: "";
+            cursor: move;
+            height: 16px;
+            width: 16px;
+            border-radius: 50%;
+            border: 1px solid var(--border-color);
+            right: 16px;
+            background-color: #ccc;
+            box-shadow: 0px 0px 8px #fff;
+        }
+    }
+}
