+++ src/components/RangeInput.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+import "./RangeInput.scss";
+
+
+function RangeInput(
+    props: Omit<React.InputHTMLAttributes<HTMLInputElement>, "type">
+): JSX.Element {
+    return (
+        <input type="range" className="range-input" {...props} />
+    )
+}
+
+
+export default RangeInput;
