+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export { default as AngleInput } from "./AngleInput";
+export { default as Container } from "./Container";
+export { default as DownloadableSVG } from "./DownloadableSVG";
