+++ src/components/properties/index.ts
@@ -1,0 +1,6 @@
+export { default as ColorGradient } from "./ColorGradient";
+export { default as LogoDimension } from "./LogoDimension";
+export { default as LogoFont } from "./LogoFont";
+export { default as LogoText } from "./LogoText";
+export { default as LogoTextOffset } from "./LogoTextOffset";
+export { default as Opacity } from "./Opacity";
