+++ src/defaultProperties.ts
@@ -1,0 +1,15 @@
+import { LogoProperty } from "./types/types";
+
+
+const defaultProperty: LogoProperty = {
+    text: "logo",
+    size: { x: 80, y: 80 },
+    font: { familyIdx: 0, variantIdx: 0, size: 20 },
+    textColor: { color1: "#FF5599", color2: "#55FFFF", angle: 45, width: 35 },
+    textOffset: { x: 0, y: 0 },
+    bgColor: { color1: "#003344", color2: "#005533", angle: 45, width: 50 },
+    bgOpacity: 100,
+}
+
+
+export default defaultProperty;
