+++ src/features/fonts/index.ts
@@ -1,0 +1,2 @@
+export { fontList, getBase64Font, getFontURL } from "./fonts";
+export { modificationAllowed, licensesWhichAllowModification } from "./license";
