--- src/index.css
@@ -1,38 +1,0 @@
-#svg-image {
-    width: min(100%, 25em);
-}
-
-.short-input {
-    width: 3em;
-}
-
-.flex {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: flex-start;
-    align-items: center;
-}
-
-td {
-    border: 1px solid #ccc;
-    padding: 3px;
-}
-
-table {
-    border: 1px solid #ccc;
-}
-
-h3 {
-    margin-bottom: 0.3em;
-}
-
-#generated-image {
-    margin: 0 1em;
-    width: max(30%, 15em);
-}
-
-#messages {
-    height: 3em;
-    overflow-y: scroll;
-}
