+++ src/types/colors.ts
@@ -1,0 +1,9 @@
+interface Gradient {
+    color1: string;
+    color2: string;
+    angle: number;
+    width: number;
+}
+
+
+export type { Gradient };
