+++ src/types/geometry.ts
@@ -1,0 +1,15 @@
+interface RectangleCorners {
+    x1: number | string;
+    x2: number | string;
+    y1: number | string;
+    y2: number | string;
+}
+
+
+interface Dimension {
+    x: number;
+    y: number;
+}
+
+
+export type { Dimension, RectangleCorners };
