+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export type { Font, FontFile, FontList } from "./fonts";
+export type { Gradient } from "./colors";
+export type { Dimension, RectangleCorners } from "./geometry";
+export type { LogoProperty, LogoPropertySetter } from "./logoData";
